+++ src/test/java/test/demo/config/Config.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package test.demo.domain;
+package test.demo.config;
 
 import java.io.File;
 import java.io.Serializable;