+++ CHANGELOG.md
@@ -214,8 +214,8 @@ public void test_Conditional() {
 - feat: add `Environment` new api to get property
 - feat: add jar-prefix file to ignore jar scanning
 - fix: ensure ExpressionFactory's instance consistent
-
-
+- fix: DataSize.parse()
+- :white_check_mark: add some test code